backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.3.0
pydantic-settings>=2.2.0
aiofiles>=23.2.0
